+++ src/InvestmentApp.Application/Common/Interfaces/IStockInfoProvider.cs
@@ -89,6 +89,6 @@ public class TradingHistorySummaryInfo
     public decimal ChangeDay { get; set; }
     public decimal ChangeWeek { get; set; }
     public decimal ChangeMonth { get; set; }
-    public decimal Change3Month { get; set; }
-    public decimal Change6Month { get; set; }
+    public decimal? Change3Month { get; set; }
+    public decimal? Change6Month { get; set; }
 }